shared/requirements.txt
# =====================================================
# PULPOAI SHARED LIBRARIES - REQUIREMENTS
# =====================================================
# Dependencias para las librerías compartidas
# =====================================================

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Data processing
numpy==1.24.3
pandas==2.0.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
